Move SamplerOutput to sequence.py
